requirements.txt
torch~=2.4.0
torchaudio
torchvision
-f https://download.pytorch.org/whl/cu121
hydra-core==1.3.2
tqdm==4.65.2
imageio
pysodmetrics

